cli: Phase C — convert verb over shared bulk-convert runner
Lifts start_bulk_convert/run_bulk_convert/scan_source_folder out of
gui/bulk_convert_dialog.rs into src/model/bulk_convert_runner.rs so
the GUI's Bulk Convert dialog and a new rb-cli convert verb both go
through the same worker. The dialog re-exports start_bulk_convert
and keeps its UI logic (review checklist, conflict modal) untouched.

rb-cli convert IN OUT --format FMT accepts a single file or a folder
of inputs, mirrors per-file progress + log to stderr, and skips
existing outputs unless --overwrite is passed.
